Feat: Support overriding entrypoint and args passed to the statefulset
Why do we need this change?
=======================
To support changing the base image to clamav upstream we need/want to enable allow modifications that will directly impact the runtime. Clamav has two init scripts that can run. one runs privledged (default) and one does not.
Default beheavior would be to use the upstream default which would be privledged, but security wise folks would want to run the unprivledged mode and we should make that possible

What effects does this change have?
=======================
Adds 2 new properties for args and command to adjust/change the way the container we run works